Improved TCB interface
The first change was explicitly setting the move constructor/assignment
operator to their default implementation, and disabling the copy
constructor/assignment operator. This was done to ensure that only
move semantics will ever be used when transferring TCB data between
data structures.

The second change was encapsulating the stack deallocation code into a
member function, freeStack. This is more expressive, easier to read and
modify if necessary, and in general is a much better practice than
manually deleting the stack by accessing a TCB's ctx directly.