Fixed freeing incremented stack ptr when cleaning up finished threads
When deleting the stack we were accessing the stack pointer of the TCB
context directly and calling delete on that. This is incorrect, however,
because the context's stack pointer is incremented whenever the TCB is
updated, resulting in a different pointer from the original one by the
time the thread finishes.

We fix this by keeping track of the original stack pointer in our TCB
and freeing that instead when we clean up the finished threads.