Extracted cleanup code into function
The cleanup code will have to run in several places, so it's best to turn
it into a function to avoid code duplication. Right now we believe we'll
have to call it directly after switching into a new context since that's
the soonest time we can free the stack of a finished thread safely.