Cleaned up comments and added asserts
For the comments, we took out some older/no longer applicable comments, changed the
wording of some comments to make them clearer, and inserted new comments about where
to acquire/relese the guard, and where to send IPIs.

We also added asserts to check if interrupts are disabled (switch invariant) wherever
possible. Some examples of where we added them are in functions that are meant to be
called in OS code, and also directly after returning from a context switch in
functions which have the possibility of resuming execution (e.g. thread::join,
mutex::lock, cv::wait).