Updated switch_to_next_or_suspend interface
Updated the function so that it now accepts nullptr. When passed nullptr as argument,
it is assumed that the caller does not want to update any context with the current
CPU registers, so we call setcontext instead of swapcontext. Passing a valid non-zero
ucontext_t pointer will result in swapcontext being used like before. This change
eliminates the need for the dummyCtx global variable, so it has been removed.