Created a test for mutex style that exposes a bug in mutex::unlock()
Currently we keep track of a mutex's owner using the raw Tcb pointer of the thread
that owns it. If a thread exits without unlocking a mutex, the following situation
can occur:

The thread that exited holding the mutex is freed. Then a different thread is created
and its Tcb is allocated at the same location as the thread which exited. This thread
can then call unlock on the mutex succesfully since its Tcb pointer is the same as
the Tcb pointer the mutex is keeping track of.

The test case added recreates the situation explained above. We will need to fix this
bug in the future. Also this bug is likely related to test case 19 on the autograder,
refer to @849 and @755 on Piazza.