Revert "Added cleanup function to handle_ipi"
This reverts commit d5fb5dbdd52b6043c3cdb55cbae4c3efdc11ca95.

This commit may have resulted in our program taking too long for test
cases 39-44, so we're reverting it. One point for why this might be true
is that handle_ipi gets called A LOT in a multiprocessor, so calling
the cleanup function inside might add unnecessary overhead. We
can't actually confirm that this is the case, however.